MA-58: Move drag drop code to interactive view
