+++ Views/NodeView.axaml
@@ -15,7 +15,7 @@
 	</Design.DataContext>
 
 	<Grid RowDefinitions="*,Auto,Auto,Auto">
-		<views:DragDropImageView Grid.Row="0" DataContext="{Binding Node}"/>
+		<views:ImageView Grid.Row="0" DataContext="{Binding Node}"/>
 		<TextBox Grid.Row="1" Text="{Binding Node.Name}" DockPanel.Dock="Bottom"/>
 		<TextBox Grid.Row="2" Text="{Binding Node.Desc}" DockPanel.Dock="Bottom"/>
 		<Expander Grid.Row="3" Header="Notes" VerticalAlignment="Top">
